More progress towards config UI, also fix crash when removing from queue

+++ src/main/java/com/lovetropics/minigames/client/screen/flex/Box.java
@@ -120,7 +120,7 @@ public final class Box {
 		return axis == Axis.X ? new Interval(this.left, this.right) : new Interval(this.top, this.bottom);
 	}
 
-	public boolean contains(int x, int y) {
+	public boolean contains(double x, double y) {
 		return x >= this.left && y >= this.top && x < this.right && y < this.bottom;
 	}
